Move lock functionality to Home button
